long type getDuration change int type，ScreenBrightness

+++ ijkplayer_coustom/src/main/java/com/daomingedu/ijkplayertest/coustomview/CustomPlayer.java
@@ -52,7 +52,7 @@ public interface CustomPlayer {
 
     void reset();
 
-    long getDuration();
+    int getDuration();
 
     long getCurrentPosition();
     void setCurrentPosition(int currentPosition);
